requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
python-dotenv